+++ cuda/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `MappedMem`, `MemProp`, `PhyMem`, `VirByte` and `VirMem` to use Device Virtual Memory;
 - Add `memcpy_d2h` to `Stream` for page-locked memory;
 - Add `index` to `Device` for device index;
+- Derive `Clone, Copy` for `Device`;
 
 ### Changed